+++ igseq/__main__.py
@@ -263,6 +263,10 @@ def _main_tree(args):
         pattern=args.set_pattern,
         lists=args.set_list,
         colors=args.set_color,
+        merge_colors=args.merge_colors,
+        figtree_opts=args.figtree_option,
+        set_pos=args.set_pos,
+        set_pos_msa=args.set_pos_msa,
         colmap=colmap,
         dry_run=args.dry_run)
 
@@ -599,6 +603,19 @@ def __setup_arg_parser():
     p_tree.add_argument("--set-color", "-C", action="append",
         help="setname=colorcode, like set1=#ff0000, to override automatic set colors. "
         "This can be given multiple times for multiple set/color pairs.")
+    p_tree.add_argument("--merge-colors", "-M", action="store_true",
+        help="merge the set colors applying to each sequence?  Otherwise (the default) apply "
+        "colors set-by-set with the last set taking precedence")
+    p_tree.add_argument("--figtree-option", "-F", action="append",
+        help="key=value to include in a figtree block for NEXUS output, like "
+        "'nodeLabels.fontSize=8'.  This can be given multiple times.")
+    p_tree.add_argument("--set-pos",
+        help="Position ranges to use to define set membership for sequences.  "
+        "If input is not an alignment, --set-pos-msa is required.")
+    p_tree.add_argument("--set-pos-msa",
+        help="alignment FASTA to use with --set-pos to define set membership.  "
+        "Only required if sets are being defined by alignment position and "
+        "input is not an alignment.")
     p_tree.set_defaults(func=_main_tree)
 
     return parser