+++ test_igseq/test_tree.py
@@ -102,6 +102,69 @@ class TestTreeEmpty(TestBase):
             tree.run_fasttree([])
 
 
+class TestTreeFigTree(TestBase):
+    """Test with optional FigTree settings in NEXUS output."""
+
+    def test_tree(self):
+        with TemporaryDirectory() as tmpdir:
+            self.redirect_streams(lambda:
+                tree.tree(
+                    [self.path/"seqs.aln.fasta"],
+                    Path(tmpdir)/"tree.nex",
+                    figtree_opts=["branchLabels.fontSize=8"]))
+            self.assertTxtsMatch(self.path/"tree.nex", Path(tmpdir)/"tree.nex")
+        # It should warn if the output isn't NEXUS
+        with TemporaryDirectory() as tmpdir:
+            with self.assertLogs(level="WARNING") as log_cm:
+                self.redirect_streams(lambda:
+                    tree.tree(
+                        [self.path/"seqs.aln.fasta"],
+                        Path(tmpdir)/"tree.tree",
+                        figtree_opts=["branchLabels.fontSize=8"]))
+            self.assertEqual(len(log_cm.output), 1,
+                "a warning should be logged for FigTree options and non-nex output")
+            self.assertTxtsMatch(self.path/"tree.tree", Path(tmpdir)/"tree.tree")
+
+
+class TestTreePos(TestBase):
+    """Test defining seq sets by what they have in an alignment."""
+
+    def test_tree(self):
+        # simple case, where the alignment is the input itself
+        with TemporaryDirectory() as tmpdir:
+            self.redirect_streams(lambda:
+                tree.tree(
+                    [self.path/"seqs.aln.fasta"],
+                    Path(tmpdir)/"tree.nex",
+                    set_pos="26"))
+            self.assertTxtsMatch(self.path/"tree.nex", Path(tmpdir)/"tree.nex")
+        # but if we're using something else as input (like an existing tree
+        # file) the alignment needs to be given separately
+        with TemporaryDirectory() as tmpdir:
+            with self.assertRaises(IgSeqError):
+                self.redirect_streams(lambda:
+                    tree.tree(
+                        [self.path/"tree.tree"],
+                        Path(tmpdir)/"tree.nex",
+                        set_pos="26"))
+        with TemporaryDirectory() as tmpdir:
+            self.redirect_streams(lambda:
+                tree.tree(
+                    [self.path/"tree.tree"],
+                    Path(tmpdir)/"tree.nex",
+                    set_pos="26", set_pos_msa=self.path/"seqs.aln.fasta"))
+
+    def test_tree_colors(self):
+        # With alignment-based sets, the set name is the sequence region
+        with TemporaryDirectory() as tmpdir:
+            self.redirect_streams(lambda:
+                tree.tree(
+                    [self.path/"seqs.aln.fasta"],
+                    Path(tmpdir)/"tree_colors.nex",
+                    set_pos="26",
+                    colors=["A=#ff0000"]))
+            self.assertTxtsMatch(self.path/"tree_colors.nex", Path(tmpdir)/"tree_colors.nex")
+
 class TestTreeMulti(TestBase):
 
     def setUp(self):
@@ -133,17 +196,28 @@ class TestTreeMulti(TestBase):
         # a special case: multiple FASTA inputs will allow implicit set
         # membership based on which files have which sequences
         with TemporaryDirectory() as tmpdir:
-            stdout, stderr = self.redirect_streams(lambda:
+            self.redirect_streams(lambda:
                 tree.tree(
                     [self.path/"seqs.fasta", self.path/"seqs2.fasta"],
                     Path(tmpdir)/"tree.tree"))
             self.assertTxtsMatch(self.path/"tree.tree", Path(tmpdir)/"tree.tree")
+        # default behavior: for overlapping set membership for any given
+        # sequence, the color from the last set is applied
         with TemporaryDirectory() as tmpdir:
-            stdout, stderr = self.redirect_streams(lambda:
+            self.redirect_streams(lambda:
                 tree.tree(
                     [self.path/"seqs.fasta", self.path/"seqs2.fasta"],
                     Path(tmpdir)/"tree.nex"))
             self.assertTxtsMatch(self.path/"tree.nex", Path(tmpdir)/"tree.nex")
+        # merge behavior: for overlapping set membership for any given
+        # sequence, the colors from all sets containing that sequence are
+        # merged
+        with TemporaryDirectory() as tmpdir:
+            self.redirect_streams(lambda:
+                tree.tree(
+                    [self.path/"seqs.fasta", self.path/"seqs2.fasta"],
+                    Path(tmpdir)/"tree_merged.nex", merge_colors=True))
+            self.assertTxtsMatch(self.path/"tree_merged.nex", Path(tmpdir)/"tree_merged.nex")
 
 
 class TestTreeMultiGaps(TestTreeMulti):
@@ -225,7 +299,6 @@ class TestColors(TestBase):
             {"A": [190, 66, 41]})
         # how about *no* sets?
         self.assertEqual(tree.make_seq_set_colors({}), {})
-
 
 
 class TestLooksAligned(TestBase):